# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yatri Travels API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List tour packages
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: minDuration
          schema:
            type: integer
        - in: query
          name: maxDuration
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"

    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a tour package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a package by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found

    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a tour package (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"

    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a tour package (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Bookings
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews
      parameters:
        - in: query
          name: packageId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Gallery
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
...
    Package:
      type: object
      required: [id, title, location, description, images, price, duration, seatsAvailable, featured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        duration:
          type: integer
        seatsAvailable:
          type: integer
        featured:
          type: boolean
        highlights:
          type: ["string", "null"]
        inclusions:
          type: ["string", "null"]
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    PackageInput:
      type: object
      required: [title, location, description, images, price, duration, seatsAvailable]
      properties:
        title:
          type: string
        location:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        duration:
          type: integer
        seatsAvailable:
          type: integer
        featured:
          type: boolean
        highlights:
          type: string
        inclusions:
          type: string
    PackageUpdate:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        duration:
          type: integer
        seatsAvailable:
          type: integer
        featured:
          type: boolean
        highlights:
          type: string
        inclusions:
          type: string
...
    Review:
      type: object
      required: [id, userId, packageId, rating, comment, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [packageId, rating, comment]
      properties:
        packageId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
    ReviewUpdate:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string

    # Gallery
...