# lib/api-zod/src/generated/types/packageUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yatri Travels API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageUpdate {
  title?: string;
  location?: string;
  description?: string;
  images?: string[];
  price?: number;
  duration?: number;
  seatsAvailable?: number;
  featured?: boolean;
  highlights?: string;
  inclusions?: string;
}